pyproject.toml
[project]
name = "langchain-langraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.13.4",
    "bs4>=0.0.2",
    "chromadb>=1.0.15",
    "colored==2.2.4",
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.0",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.28",
    "langchain-text-splitters>=0.3.9",
    "langgraph>=0.6.4",
    "langserve>=0.3.1",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "sse-starlette>=3.0.2",
    "streamlit>=1.48.0",
    "uvicorn>=0.35.0",
    "wikipedia>=1.4.0",
]
